updated- added a functions folder for the different functions for each page
